change p to uppercase P
from the spec:
> To avoid confusion, we will use lower-case p to mean the former (0-1) and upper-case P whenever we mean the latter (0-255).

Since this implementation is only P at the moment, I figure the argument
should reflect that. Supporting p as a type would be desirable, but
there are issues with that:
* converting between p and P can lose precision, making round tripping
harder to test (though, tbh, the difference is probably negligible..)
* how to represent p? as a float 0-1? a UInt8 with range of 0-100?
Easier really to just use P, leave converting as an exercise for the
user. Not like this is very end-user facing :), and also it's probably
a one time thing, so it's fine.