convert to throwing
So, I think I like throwing better than return nil. Reason being, if you
want to treat it as a return nil, you can just `try?` and it is the same
semantics. However, if you do get an error, at least you know *what*
happened, instead of just getting a nil. And if you want you can do away
with a lot of guard lets or if lets and just catch the whole thing if
you want.