Use a less fragile method for return frames
*currently* the only way the designated initializer can fail is if the
port number is invalid, which it can't be in this case. However, that
may change (doubtful, but defensive programming here). So instead of
calling another initializer and ignoring its possible failures, let's
just make ourselves by ourselves.

This also is fine because one of my branches turns this into a struct
type rather than a class, so there won't be any convenience initializers
anyways